CEIA-LLMIAG/ClaseVI/codigo/requirements.txt
transformers
Pillow
torch
groq
streamlit
langchain
langchain_core
langchain_groq
numpy
pinecone
sentence-transformers
openai
datasets
pinecone-text
tqdm
IPython
python-dotenv
